requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
ls-qa-api==1.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyMySQL==1.0.2
pytest==7.2.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.14
WooCommerce==3.0.0
